Abstracted 'utils.py' as a module containing py files for flask (app), audio, db, gcs, and factory logic
