+++ server/tests/integration/test_ping.py
@@ -1,6 +1,7 @@
 import pytest
-from tests.utils import drop_all_collections, get_test_app
 
+from tests.utils.test_app import get_test_app
+from tests.utils.test_db import drop_all_collections
 
 @pytest.fixture
 def app():
